-Working with 2D, problem: need to understand
--What i have changed--

-In atlas training:
I did an interpolation before the torch.cat and commented a redundant pooling layer --> forward pass in Unet

-In whole training
The Unet was already fixed and the classifier network i had to put an adaptive pooling layer that works with small spatial sizes and another interpolation to match the sizes.

Big question: Do this really works? We dont know, i need to understand now the code and see what is happening